Move Login and UserSignup into auth/ directory
